Web interface: Change page title from "Web of Trust" to "Community"; allow translating
For the Freenet default theme, this is the very large title which the
web interface will display above the whole Freenet interface.

It is better to use "Community" because that is also the menu name which
WOT is accessed through.

In addition, this fixes the string to not be hardcoded anymore but to
use the L10n framework so it can be translated.