WebPageImpl: Add member variable which holds the logged in OwnIdentity
This will be used by the following commits to:
- replace code duplication in all child classes since the callers of the
affected constructor all already query the OwnIdentity on their own.
- obtain the nickname from it so the WebPageImpl constructor can put it
into the page title to have a title like "Logged in as: $nickname"

Notice: The added throws declaration of "UnknownIdentityException" to
one constructor does not need any adaption of the callers of that
constructor - they already throw it themselves because they query the
identity on their own.