KnownIdentitiesPage: Support sort by edition
English and German l10n included since I am German.

Notice:
This needs no further adaption to work together with the sort function
in class WebOfTrust (which supports sorting by edition since the
previous commit) because of that line of the KnownIdentitiesPage:
  WebOfTrust.SortOrder sortInstruction
    = WebOfTrust.SortOrder.valueOf("By" + sortBy + sortType);

I.e. the value of the WebOfTrust.SortOrder enum is automatically
determined from the string which we receive by HTTP (and which we
produced in the HTML form from the local enum of the
KnownIdentitiesPage).