IdentityFileDiskQueue: Implement add()
We synchronize on Java level instead of using file locking because Java
FileLock documentation says:
"File locks are held on behalf of the entire Java virtual machine. They
are not suitable for controlling access to a file by multiple threads
within the same virtual machine."
... but we need to guard against multi threading since the
IdentityFetcher will call add() concurrently. (This is because fred
will deliver files concurrently to it.)

Source:
http://docs.oracle.com/javase/6/docs/api/java/nio/channels/FileLock.html

The deprecated class Closer is used because fred does not allow plugins
to require Java 7 yet.