IdentityFileDiskQueue: Fix file format to contain FreenetURI
We need to store the FreenetURI of the file inside it since the identity
XML does not and cannot contain it: It is an USK and thus might change
during insert.
We also cannot store the URI in the filename since deduplication then
wouldn't work anymore: The filename needs to be always the same for
every file from a single identity, so it may not contain the USK
edition. It may only contain the identity ID, as it does now.

The serialization functions write/read() are copied from
Persistent.serialize() / deserialize().
I think "write" and "read" are better names since they are more widely
known words, especially when being used together with files as we do
here.