IdentityFileDiskQueue: Move statistics variables to new class IdentityFileQueueStatistics
This was done automatically using Eclipse, it should be complete.

Motivation:
The future goal behind the statistics is to add them to the web
interface.
This will require adding getters for them. If the web interface used a
getter for each number, then the numbers would not be coherent.
Thus, we would have to require the web interface to stay
synchronized(IdentityFileDiskQueue.this) while calling all individual
getters.

To prevent the synchronized() block in the web UI, we will
instead encapsulate it in a IdentityFileDiskQueue function "public
synchronized  IdentityFileQueueStatistics getStatistics()" which creates
a coherent clone of our stats object and returns it.