IdentityFetcher: Enqueue files to IdentityFileQueue, move processing code to IdentityFileProcessor
The synchronized(mWoT) can be dropped in the code moved from
IdentityFetcher to IdentityFileProcessor since
XMLTransformer.importIdentity() does it on its own.
It was only being done in IdentityFetcher to preserve the locking order
since it takes the lock on itself before calling importIdentity().

The "mRequests.containsKey(identityID)" check in IdentityFetcher can be
dropped as well: The fact that we queue identity files means that they
will be processed a long time, possibly minutes, after we queue them.
This means that the checks result can likely be invalid then.
It was a check with false positives anyway, and was being done here only
to avoid doing the costly locking in XMLTransformer.importIdentity().
Removing the check allows the whole function to not require the
IdentityFetcher lock, which is good because the lock can be expenisve to
acquire due to the lengthy run() of IdentityFetcher.
The check does not have to be re-introduced in IdentityFileProcessor
since XMLTransformer.importIdentity() does it anyway, and since we
shouldn't mess around with IdentityFetcher's internals in
IdentityFileProcessor. A check which can yield false-positives is very
ugly anyway.

The counting of fetched files can be removed in IdentityFetcher since
IdentityFileDiskQueue counts them already.
The timing measurement is removed too since will be done in
IdentityFileProcessor. A FIXME for this was added already in a previous
commit
Its good to remove both those statistics computations since they would
have required the expensive IdentityFetcher lock.

The catch(Exception e) from IdentityFetcher can be changed to
catch(RuntimeException) in IdentityFileProcessor since
bucket.getInputStream() which causes the non-RuntimeException is not
done there.