IdentityFile: Fix checksum being cached in serialized data, again
The type change of mURI from FreenetURI to String of the previous commit
is intentionally not reverted:
Serializing String only should make this a bit faster and less dependent
upon serialization format changes of FreenetURI, which is code of a
different project.
Using String instead of FreenetURI also is a good preparation for using
a more simple, faster file format than Java serialization.