Add class Base32 from Freemail
Based upon Freemail commit ccda02a78be0dc92fefb2c6245521684eb05d610.
The only modification is changing the package.

Will be used for encoding IdentityIDs in a way which can safely be used
as a filename on multiple platforms.
I would have used Base36, but we don't seem to have an implementation in
any of the Freenet repositories yet.
And Base32 has some advantages:
- It doesn't contain characters which are easy to mix up by humans, so
we might someday even use it in the UI. Attacking identities by
producing a public key with a colliding prefix is a cheap attack, and
it can be made more difficult by preventing character mixups.
- Freemail already uses Base32, so it might make sense to use it as well
in WOT for a unified format of Identity IDs.