IdentityFileQueueTest: Remove resolved FIXME
testByComparingResultsOfTwoImplementations() does implicitly test
whether deduplication does not deduplicate too much, as setUp() creates
lots of test files with old editions which should get deduplicated.

The other part of the FIXME is downgraded to TODO: If deduplication
doesn't deduplicate when it could, that will only cause WOT to be
slower, it will not cause wrong computation results.