Update German translation, review it, improve usability
Thanks a lot to the translators on Transifex which had completed WOT's
little amount of German translation.
I've reviewed all of the new and existing strings, and tried to improve
a lot of them.
As a consequence, the English l10n should be reviewed against the German
one soon to backport the improvements. A bug was filed for this at
https://bugs.freenetproject.org/view.php?id=6569

List of improvements - described in English since many of them apply to
the English language as well:
- Replace "insert URI" with "secret key" (requested at
https://bugs.freenetproject.org/view.php?id=6214)
- Replace "request URI" with "public address" or simply "address"
- Prefix back/continue buttons with "<" / ">" as is common in wizard UIs
- Replace "context" with "service", see
https://bugs.freenetproject.org/view.php?id=6568
- Replace "introduction puzzle" with "CAPTCHA" as this is a word which
is widely used for the same thing on the web
- Replace "score" with "computed trust". See
https://bugs.freenetproject.org/view.php?id=6570. Also explain better
that those computed values are the fallback if the user has not
specified a trust on his own.
- Keep menu at the English word "Community" as it is widely
known in Germany
- Fix various mistranslations
- Reduce length of sentences
- Restructure and rephrase many sentences for simplicity
- This list is incomplete since some subtle aspects are very difficult
to explain and thus beyond the scope of a commit message. If you adapt
the English l10n to improve what the German improved, please do this by
comparing the whole of both l10ns.