WebOfTrust.updateScoresWithoutCommit(): Fix AssertionError
The removed assert would fail in unit tests related to
restoreOwnIdentity()/deleteOwnIdentity() because at the point where it
is executed, duplicates of the affected identity exist in the database
and computeAllScoresWithoutCommit() does not work with duplicates in the
db.

The end of the function does contain this assert() already anyway - and
in a proper way: Only if mTrustListImportInProgress is false. That
boolean is used by the restoreOwnIdentity()/deleteOwnIdentity() to
prevent computeAllScoresWithoutCommit() from being executed at this
point because it will fail as it did in the removed assert().
If the boolean is true, then an assert(computeAllScoresWithoutCommit())
will also be executed elsewhere - at the end of the trust list import
when the duplicate identities have been removed and thus where
computeAllScoresWithoutCommit() can work properly.