WebOfTrust.computeScoresAfterDistrust...(): Use memory instead of database; fix wrong event notifications
We're not sure yet how many Scores are affected in average, so we may
assume that their amount is small and thus keep them in memory instead
of storing them in the database.

This also fixes creation of event-notifications with a wrong "before"
value:
For scores which were newly created, the "before" value was non-null,
and now is fixed to be null.

The boolean Score.mRankOudated which is made obsolete by this
commit was not actually committed yet, so it is not removed by this
commit.