WebOfTrust.computeRankFromScratch(): Only mark vertex as seen if we actually processed it
This should not yield any different result IMHO:
The PriorityQueue returns identities sorted by rank.
As soon as it returns identities with the maximum rank, it will not ever
return any ones with a rank below maximum: Identities with maximum rank
cannot give a lower rank to other identities, and there cannot be any
identities with a rank higher than the maximum.
So it doesn't matter whether an identity will now be allowed for
processing its neighbour ranks due to the change of the seen-policy, as
it cannot yield any lower ranks.

However, I am changing this nevertheless even though the results are the
same
- It improves readability
- It improves performance as we won't call seen.add() as often