WebOfTrust: Add updateCapacitiesAfterDistrustWithoutCommit()
This is an optimized version of the existing code in
updateScoresAfterDistrustWithoutCommit() which detects which Score's
capacties are outdated.

The old code would blindly mark all Score's capacities as outdated which
are reachable from the distrusted Identity, no matter whether they were
touched by an actually outdated rank.

The new code will only process capacities for Scores of which the rank
changed, as the capacity is computed from the rank.
Also, it will only yield Score objects whose capacities actually did
change.