WebOfTrust.computeAllScores...(): Validate IdentityFetcher state
If the "should I fetch this Identity?" state of the IdentityFetcher is
false, correct it.
Also, return false in that case : The boolean return value of
computeAllScores...() is used in assert()s to test the new incremental
score computation code. If the code doesn't properly deal with the
IdentityFetcher, that is a bug.

Notice: This should cause unit tests to fail since the
incremental score computation code does not deal with the
IdentityFetcher yet. This will be implemented in following commits.