WebOfTrust.updateScoresAfterDistrust...(): Fix failure of WoTTest.testRemoveTrust3()
The bug here was that when updateRanksAfterDistrustWithoutCommit()
detected a rank changing from >0 to "none", it wouldn't include that
change in the return value.

As a consequence, updateCapacitiesAfterDistrustWithoutCommit() wouldn't
determine that the capacity changed, and wouldn't return that to the
caller updateScoresAfterDistrust...().

Then updateScoresAfterDistrust...() wouldn't update the Score values
in whose computation the removed capacity was involved.

Notice: The test will still fail if you don't comment out code as
explained in the previous commit
9da69f6bfe1b21cbab778b121a50a9d2a8f7bd2f.
(It wasn't possible to sort the commits in a different order since what
this commit here adds is also the preparation to fix the other test
failure. This fix here is more important as it is less obvious, so I
chose to do it first.)