WoTTest.testRefetchDueToCapacityChange(): Fix false failure
For currently unknown reasons, even the reference implementation of
Score computation (computeAllScoresWithoutCommit()) will mark an edition for
refetch, i.e. set it as NotFetched, upon change of shouldFetchIdentity()
from false to true. I.e. it will also refetch an identity if the
capacity changed from "nothing" to 0.
This is what the initial setTrust() in this test function does, so it
would cause the edition fetch state to be set to NotFetched.

Thus, we must mark the edition as Fetched *after* the first setTrust().

The weirdness of the reference implementation was also added to the new
implementation at 1e13581f2656f64cef6e7c58d4b2829415137059.

The unit test should succeed for the new implementation now.