IdentityFetcher.DEBUG__NETWORK_DUMP_MODE: Fix not fetching anything
subscribeContent() with runBackgroundFetch==false as the old code did it
would not fetch any data if the user did not manually browse to USKs on
FProxy.
In other words: The background fetch is the active thing which fetches
USKs.

The goal behind the old code was to disable the *sparse* fetching which
subscribeContent()'s JavaDoc for runBackgroundFetch talks about; not to
disable the whole USK fetching.
So we only disable the sparse fetching now.