WOTUtil: Fix "Damaged database, exiting!"
The actual fix is which this commit does is this:
IdentityFetcher.getShouldFetchState(): Fix to work with
constructor WebOfTrust(String).

When using the constructor WebOfTrust(String), an IdentityFetcher is
created which has no USKManager.
With the old code, this would happen:
- IdentityFetcher.start() would try to fetch all Identitys
- fetch(USK) (which this commit modifies) would throw because there is
no USKManager
- Since fetch(USK) failed for all Identities, the mRequests map would
not indicate any Identity being fetched.
- computeAllScoresWithoutCommit(), which the WOTUtil calls using
verifyAndCorrectStoredScores(), would detect that Identitys which should
be fetched are not being fetched, and return false.
- The false return value would cause WOTUtil to complain about the
database being damaged.

The fix is to:
- change fetch(USK) to not throw if there is no USKManager, but return
null.
- This will make the calling code put a key in the mRequests map with
null value.
- The fact that there is a key for the identity in mRequests will be
enough to cause IdentityFetcher.getShouldFetchState() to return "true"
to indicate  that the identity is being fetched.
- And getShouldFetchState() returning true will cause
computeAllScoresWithoutCommit() to be satisfied and return true
- And thus verifyAndCorrectStoredScores() will also return true, and
WOTUtil will be happy.