WebOfTrust: Fix database corruption when trying to load databases which are newer than the current WOT plugin
When trying to determine wheter a database is too new for us, we used to
open it in read-*write* mode.
In the situation where in comparison to the currently running WOT code
the newer database contained changes to the class structure such as
adding new member variables to classes, this would cause db4o to delete
those changes from the new database: If the currently running code's
classes does not contain them, db4o will erase the fields from the
database.

When the user then continued to try to load the database with a recent
WOT version again, the new version would run into NullPointerExceptions
for example due to the deleted fields.

This commit fixes this problem by opening the database in read-only mode
when trying to determine whether it is too new for us.

Also the commit makes the error message in the case of trying to load a
too-new database more self-explanatory.