WebOfTrust.computeRankFromScratch_Caching(): Use cache to speed up itself
This not only causes the function to return immediately if the cache
contains the requested rank, but also changes the UCS (= SPSP) algorithm
to consume the cache, and thus should be a major speedup.

Notice: This contains many very slow asserts which should be commented
out. Committing this nevertheless already since I've been working on the
code for quite a few hours and it would be good to do a backup now: The
RankComputationTest unit test succeeds now.