WebOfTrust.computeRankFromScratch_Caching(): Fix unit test failures
updateCacheWithMyPath() had the bug of not properly handling the case
where the second processed vertex is an Integer.MAX_VALUE rank. It would
believe there was a skipped chain of vertices (see JavaDoc) because
"v.rank - v.previous.rank != 1", but there was none actually. MAX_VALUE
is just set on all vertices in that case by the code which runs before
this function.

Now, "v.rank - v.previous.rank > 1" is used to detect a skip instead
of != 1. This ensures that we do not wrongly detect a skip if there was
none in the case where all vertexes have a rank of Integer.MAX_VALUE.

Also, rank computation using "reversedRank = v.rank - v.previous.rank"
in the case of a real skip was wrong somehow as well I think. Don't
remember the details.
It is fixed by loading the rank of the second vertex from the
cache instead of computing it by the rank difference of first and
second.

Also, the skip detection code is pulled out of the main loop for
performance: It can only ever have one hit anyway.