ScoreTest: Ensure correctness of database in terms of WebOfTrust.verifyAndCorrectStoredScores()
This shouldn't change the behavior of the test.
It will however ensure that
    assertTrue(mWoT.verifyAndCorrectStoredScores());
would succeed if it was being called after the tests.
This is necessary because the following commit will make the parent
class do that.

The correctness of the Score database is now ensured by:
- Instead of manually creating OwnIdentity objects, we now use
WebOfTrust.createOwnIdentity(): This ensures that
WebOfTrust.initTrustTreeWithoutCommit() is called, which creates
certain special Score objects.
- Instead of only storing a manually created Score object, we now also
store a Trust object which would result in creation of the given
Score object if the Trust object was being created through
WebOfTrust.setTrust(). In other words, we create a Trust object which is
coherent with the Score we want to create.