IdentifierHashSet.add(): Fix modifying set when it should not
The main purpose of this class is to be robust against clones of
Persistent objects which have a different state.
Thus accepting them in add() like the old code did is NOT an option, it
would be exactly what we want to avoid with this class.

Symptoms would be:

    [junit] Testcase: testIterator took 2.795 sec
    [junit]     FAILED
    [junit] null
    [junit] junit.framework.AssertionFailedError
    [junit]     at
plugins.WebOfTrust.util.IdentifierHashSetTest.testIterator(IdentifierHashSetTest.java:324)