IdentifierHashSetTest: Improve test data set
The one goal with the duplicates is to have clones for which equals()
is false when comparing to the original. An unmodified clone would have
equals() == true, which is not what we want. It could even make tests
falsely succeed because we batch process the clones and thus a matching
clone could hide the effect of a non-matching one.
So we don't add unmodified clones anymore now.

Also, remove the commented out code of adding the original element to
the clones even, as thats an even more obvious violation of the above
and thus shouldn't need an explanation.