IdentifierHashSet.equals(): Fix to comply with Object.equals() specification
Specification demands that we should return false for obj == null, not
throw.
Arguably, this includes the case where the obj is not an
IdentifierHashSet, so return false there as well. As a middle path,
assert(false) then.

Fixes failure of IdentifierHashSetTest.testEqualsObject().
All tests of the class should succeed now.