l10n: Update existing translations from Transifex to remove/update strings which changed in English; fix web interface to handle removed trailing whitespace
This branch changed the English l10n reference translation a lot. Thus,
many of the translated languages need the outdated strings removed. As I
have no tool to do that automatically, this commit instead pulls the
translations from Transifex to get rid of outdated ones. Since the
translators already translated a lot of strings, this also ships updates
to the translations.

New languages are not added yet, they will ship in a following commit.

Notice: Transifex has the habit of removing trailing whitespace from
l10n strings. The web interface used them as part of doing:
    showStuff("Some text: " + someVariable);
Thus, this commit fixes that code to not expect the l10n to contain a
space after the ':'.