WebOfTrust.computeRankFromScratch(): Improve readability
IdentityHashSet<Identity> did its job correctly there, as multiple db4o
queries for the same Identity will return the same Identity objects.
But "Identity"HashSet means that it compares object identity, it isn't
related to WoT's concept of class Identity.

Thus, better replace it with IdentifierHashSet, which serves the same
purpose but has a less confusing name.
The fact that it is slower doesn't matter here, as this function merely
is a reference implementation, it isn't used in practice except for unit
tests.

The comment which is removed isn't needed anymore since the JavaDoc of
IdentifierHashSet is supposed to explain the fact which the commit
mentioned.