Add xor's WoT bachelor thesis as developer's manual
Hey folks,

since it did pass examination by the university, hereby I provide you
with my bachelor's thesis of computer science.
My past year's Freenet work of improving the core algorithm of WoT had
been recycled into it.

Further, since the thesis aims to require no prior Freenet-related
knowledge, it provides a pretty thorough explanation of how WoT works.
Thus, you may consider it as the WoT core developer's manual.

Besides getting to know how WoT works, it would be of scientific benefit
for the project if you do read it:
The end of the thesis describes how the algorithm might be further
improved by investigating what can be considered as a whole class of
algorithms. I have not heard about such a class of algorithms being
identified and named by science yet. But this might be merely due to
lack of my knowledge.
So I hope that someone knows if this class of algorithms has already
been discovered and analyzed by other people. This could help us a lot
to further improve WoT with algorithms which are specially crafted to be
fast at this type of problem.

Feedback for improvement of the document is welcome:
Besides some layout changes (see [1]) the document is still identical to
what I handed in at university.
Thus, it may contain many stupid mistakes caused by pre-deadline
caffeine intoxication.

I can also provide printed copies, university wants something around 30
EUR for printing.

Greetings and thanks for reading!

[1] Changes since the original version 1.1, which I had mailed to Ian
and Steve already:
- The main text body has not been changed, so you don't need to re-read
it!
- The title page has been improved to contain the Freenet logo and
provide a subtitle.
- The "declaration of $boring_legal_stuff_for_university" page has been
removed.
- The LaTeX stuff which does line wrapping has been configured to allow
wrapping URLs, which fixes a glitch in the layout of the bibliography.