WoTTest.test_RestoreOwnIdentity_DeleteOwnIdentity_Chained(): Fix dangerous use of HashSet
Bugs in the code which this tests could cause duplicates of Identities /
Trusts/ Scores to exist.
Adding them to the HashSet like the previous code did could cause the
HashSet to deduplicate them.
This would cause the test to not detect the duplicates as bugs.

Thus, this commit fixes the code to use getters which check for
duplicates.