SubscriptionManagerFCPTest: Fix dangerous use of HashSet
Bugs could cause duplicates of Identities / Trusts/ Scores to exist.
Adding them to the HashSets like the previous code did could cause the
HashSets to deduplicate them.
This would cause the test to not detect the duplicates as bugs.

Thus, this commit fixes the code to use getters which check for
duplicates; and to check for duplicates on its own HashSets.