Revert "Merge branch 'issue-0006847-IdentityFile-dumper' into next"
This reverts commit a6be1937dab1a5ac33dcffb9cd6ef589c5d654b7, reversing
changes made to 5860fa6c2437e85a994c2fec4b470a2f53e63d45.

The reverted code was intended to provide functionality to dump the
non-human-readable file format of IdentityFile in a human readable way.
This is reverted because branch
    issue-0006846-human-readable-IdentityFile
changes the file format of IdentityFile to be human readable.

Conflicts:
	src/plugins/WebOfTrust/IdentityFile.java