Eclipse builder: Compile for Java 7
fred is on Java 7 since build 1473, so we can use it now.

Notice: The other settings are automatically added by Eclipse if I
use "Enable project specific settings" in the compiler settings, which I
have to do to be able to chose Java 7.
I've left them add their defaults though.