newHashSetFromUniqueObjects(): Rename to listToSetWithDuplicateCheck()
The duplicate check is a central feature of the implementation which
calling code will explicitely need to fulfill its testing duties.
I.e. the function aims at allowing calling code to specify collections
of *non*-unique objects; it desires to be capable to detect this to
consider it as a test failure.

Thus, if the function only implemented it, but did not externally
promise it from its function name (or JavaDoc), it is possible that
future developers remove this functionality from it.

So by making this feature part of the function name, we ensure that this
feature stays in the function.