IntroductionPuzzleStoreTest.testGetUninsertedOwnPuzzlesByInserter(): Fix dangerous use of HashSet
Bugs could cause duplicates of IntroductionPuzzles to exist.
Adding them to the HashSets like the previous code did could cause the
HashSets to deduplicate them.
This would cause the test to not detect the duplicates as bugs.

Thus, this commit fixes the code to use listToSetWithDuplicateCheck().

Bonus changes:
The "final" attribute is removed because the noise it adds to the code
is not justified given the low size of the function.
The removed import was not used.