English l10n: Fix leftover usage of "score" to be "trust value" instead
The said strings were developed on a branch in parallel to the
development of the English l10n revamp branch.
The revamp changed all occurences of "score" to "computed trust" or
"trust value".
When the branch was merged with the English l10n revamp branch, it was
forgotten to adapt them. See commit
5146a00a3f270a1284b5d56c2aab348e414ee0ed.

This commit fixes them to be similar to the lines right next to them.