WebOfTrust.deleteDuplicateObjects(): Improve performance
Yes, this is ugly.
But:
1) This function is supposed to be removed long-term anyway:
Once we move to a proper database we shall configure the database to
prevent it from even allowing duplicates to be created in the first
place.
2) The function is only called if DEBUG logging is enabled, not during
normal operation.

Notice: recomputeScores is renamed to anythingChanged as a safeguard
against this in the future being changed to even more violate the
programming best-tractice that finishTrustListImport() should be
considered as a black box.
Having the new name ensures that adding other stuff than score
computation to finishTrustListImport() doesn't break together with the
behavior of this function here: Noticing developers that they should
call finishTrustListImport() as long as they changed *anything* to the
database will ensure their changes to this function are independent of
changes at finishtrustListImport().