IdentityDownloaderController: Implement start() / terminate()
I first considered starting/stopping them in a thread each so we get
concurrent startup. Unfortunately with Java7 this still winds up to be
something like a whole screen of code. (Java 8 will reduce this a lot
thanks to lambda expressions)

And: The new implementations this branch is working on shall be
significantly faster than the old class IdentityFetcher. Thus, their
startup/shutdown will likely be so fast that we don't need to deal with
parallelizing.