IntroductionServer: Continue to wire in IdentityDownloaderController as replacement for IdentityFetcher...
... by finishing it for class IntroductionServer.

Interestingly, the class requested the IdentityFetcher as parameter to
its constructor even though it didn't need it anyway. So we can just
remove it.

(This was likely the case because it was originally planned to have the
IntroductionServer make the fetcher start fetching a remote identity
once it has become trustworthy by solving a puzzle. That's not needed
though: Solving a captcha will cause the solver to receive a trust
value, and receiving a trust value makes the WebOfTrust tell the fetcher
to start fetching).