IdentityFetcher: Continue to wire in IdentityDownloaderController as replacement for IdentityFetcher...
... by making IdentityFetcher lock upon that instead of upon itself.

This ensures we don't have to change the rest of WoT to lock upon
IdentityDownloaderFast *and* IdentityDownloaderSlow once they replace
IdentityFetcher by changing
IdentityDownloaderController.USE_LEGACY_REFERENCE_IMPLEMENTATION to
false.
I.e. those two classes will also lock upon the
IdentityDownloaderController so it doesn't matter which implementation
of IdentityDownloader is running - be it IdentityFetcher or the new
ones. The locking will be independent of the implementation by having
IdentityDownloaderController be the central lock.