IdentityFetcher: Finish wiring-in IdentityDownloaderController as replacement for IdentityFetcher...
... by deprecating IdentityFetcher.

The code should compile and execute now, and produce the same results as
before IdentityDownloaderController was introduced by this branch:
IdentityDownloaderController should start the legacy class
IdentityFetcher and pass through all calls it receives to the
IdentityFetcher.

So effectively, this branch now provides:
- an abstraction of IdentityFetcher by extracting the interface
IdentityDownloader from it.
- IdentityDownloaderController as a proxy for all calls which previously
were aimed at IdentityFetcher. This allows IdentityDownloaderController
to in the future be configured to use the upcoming new implementations
of IdentityDownloader to replace IdentityFetcher.