IdentityDownloader: Pass more information about edition hints
Previously, hints were stored in class Identity - and only the highest
received hint from anyone was stored. It also wasn't stored who gave the
hint.
This allowed malicious identities to give us too high hints to editions
which don't actually exist to DoS the hint subsystem.

To fix that, this commit changes the interface IdentityDownloader to
provide the implementations with information about who gave each hint
and what the specific hint is.
This will allow the implementations to try multiple hints if one is
bogus, and to punish Identities which provide too many bogus hints.

The classes which feed the IdentityDownloader implementations with the
event callbacks are also adapted to actually feed it with the new
information.
The existing implementation Identityfetcher is not adapted to use it
yet: I'd like to keep the class as is for the purpose of having
something to test the upcoming new implementations against.