XMLTransformer/IdentityDownloader: Accept more edition hints; pass more information about them.
This ships both refactoring and algorithmic changes:

1. Algorithmic changes:
Most notably, the decision of which hints to accept/refuse is relaxed:
Previously, they were only accepted if the giver has a positive Score
value.
Now they're also accepted if the giver merely has a positive
Score capacity (= has a right to have his Trust values be included in
the Score value computations and particularly in the decision of which
identities to download).
This is to ensure Score computation continues to be "stable" considering
the fact that Identity downloading will rely very much on edition hints
as soon as we switch from IdentityFetcher to IdentityDownloaderFast and
IdentityDownloaderSlow.
"Stable" means that the results of Score computation should be
independent from order of Trust value import. Given that we won't
subscribe to all USKs anymore and download some  identities only by
hints, we need to extend the concept of "stability" to the downloading
itself:
If we refuse too many hints and thus don't download some identities,
input for Score computation might be lacking depending on which
identities were downloaded first.
So to ensure stability we will now also accept hints if the giver has a
capacity.
For a more detailled explanation in different words, see the large
comment added to class XMLTransformer.

Notice: This includes the posibility that we download more "spam"
hints, i.e. hints to editions which don't actually exist. This
posibility has always existed, and a fix has been and is still required.
This branch will ship it as well: The new IdentityDownloaderSlow
shall track the percentage of wrong hints for each identity and punish
identities if too many of their hints are bogus. Also see below...

2. Refactoring:
Previously, hints were stored in class Identity - and only the highest
received hint from anyone was stored. It also wasn't stored who gave the
hint.
This allowed malicious identities to give us too high hints to editions
which don't actually exist to DoS the hint subsystem.

To fix that, this commit changes the interface IdentityDownloader to
provide the implementations with information about who gave each hint
and what the specific hint is.
This will allow the implementations to try multiple hints if one is
bogus, and to punish Identities which provide too many bogus hints.

The classes which feed the IdentityDownloader implementations with the
event callbacks are also adapted to actually feed it with the new
information.
The existing implementation IdentityFetcher is not adapted to use it
yet: I'd like to keep the class as is for the purpose of having
something to test the upcoming new implementations against.