IdentityDownloader*: Rename function to match new purpose
"Update" is replaced with "New": It doesn't fit anymore as this branch
aims to allow *multiple* hints for a given Identity to be stored. See
JavaDoc of class EditionHint.

This has been automated with Eclipe.
Formatting (= removing linebreaks which are not needed anymore thanks to
the shorter name) will be manually corrected in the neext commit which
is planned to change the signature of that function.