IdentityTest.testSetEdition(): Replace deprecated onFetched() with onFetchedAndParsedSuccessfully()
- The first call is moved some lines upwards because the replacement
function affects Identity.getEdition() and others so we want to call it
before checking the value of them.

- Typically onFetched() was used in combination with the
also depreacted setEdition() so  normally we would replace the whole of
the setEdition()/onFetched() combination with
onFetchedAndParsedSuccessfully(), but as this unit test *is* about
setEdition() we don't replace that one. This still allows us to get rid
of onFetched() as it won't be called by anything after this commit.
And ths whole unit test function will disappear anyway once all callers
of setEdition() are gone and setEdition() is removed so we don't need to
get rid of its usage here.