XMLTransformer: Fix to not accept an edition which we already processed; log as error if it happens
Every edition should only be downloaded and parsed ONCE (unless
Identity.markForRefetch() was used, which marks the edition as
FetchState.NotFetched so it would appear as not having been parsed yet
and thus we can ignore this case here).

This bug was possible because getEdition() will also return the very
same edition if getCurrentEditionFetchState() is FetchState.Fetched, so
we'd  process it twice without causing any ERROR logging about having
received the edition twice.

Also replace usage of the deprecated getEdition() with the new
getLastFetchedEdition().