Identity: Remove deprecated onParsingFailed()
The previous comments have adapted all callers to use the new
onFetchedAndParsingFailed() instead.

The bad thing about this function was:
When we fetched an edition and parsing failed, we need not only to mark
it as parsing failed by calling onParsingFailed(), but also store the
fetched edition number. This had to be done separately with
setEdition(). So for one event - a new edition was fetched - developers
had to always call two functions. This is prone to forgetting about
calling one of them.
The new onFetchedAndParsingFailed() fixes that by consuming the
edition number as parameter and doing both things at once.