OwnIdentityTest.testClone(): Fix bogus failure introduced by previous commit
The previous commit wanted to set up the Date value of
getLastFetchedDate() to be before the value of getLastInsertDate().

But it called onInserted() *after* onFetchedAndParsedSuccessfully()
while forgetting about the fact that onInserted() not only updates the
LastInsertDate but also the LastFetchedDate, so it destroyed the order
of the Dates it wanted to set up.

This commit fixes that by swapping the expected order of LastFetchedDate
and LastInsertDate.

The fixed failure looked like this:

Testcase: testClone took 5.921 sec
    FAILED
junit.framework.AssertionFailedError
    at plugins.WebOfTrust.OwnIdentityTest.testClone(OwnIdentityTest.java:98)